resume_parser/requirements.txt
Flask
Flask-MySQLdb
PyMuPDF  # for PDF parsing
python-docx  # for DOCX parsing
pytesseract
scikit-learn
spacy